Move useWindowSize hook to its own file
