+++ public/content/lessons/2021/newtons-fractal/newtons-fractal-interactive/index.js
@@ -1,4 +1,5 @@
-import { useCallback, useEffect, useState } from "react";
+import { useState } from "react";
+import useWindowSize from "../../../../../../util/useWindowSize";
 import GraphWindow, {
   GraphPoint,
   GraphLines,
@@ -210,29 +211,6 @@ function multiply(...nums) {
       ...nums.slice(2),
     ]
   );
-}
-
-function useWindowSize() {
-  const [size, setSize] = useState({
-    width: window.innerWidth,
-    height: window.innerHeight,
-  });
-
-  const updateSize = useCallback(() => {
-    setSize({
-      width: window.innerWidth,
-      height: window.innerHeight,
-    });
-  }, []);
-
-  useEffect(() => {
-    window.addEventListener("resize", updateSize);
-    return () => {
-      window.removeEventListener("resize", updateSize);
-    };
-  }, [updateSize]);
-
-  return size;
 }
 
 export const vertex = `#version 300 es
