+++ util/useWindowSize.js
@@ -1,0 +1,24 @@
+import { useCallback, useEffect, useState } from "react";
+
+export default function useWindowSize() {
+  const [size, setSize] = useState({
+    width: window.innerWidth,
+    height: window.innerHeight,
+  });
+
+  const updateSize = useCallback(() => {
+    setSize({
+      width: window.innerWidth,
+      height: window.innerHeight,
+    });
+  }, []);
+
+  useEffect(() => {
+    window.addEventListener("resize", updateSize);
+    return () => {
+      window.removeEventListener("resize", updateSize);
+    };
+  }, [updateSize]);
+
+  return size;
+}
